pkg/api/core: move pkg/api/pod
