+++ pkg/kubectl/.import-restrictions
@@ -5,7 +5,7 @@
                 "k8s.io/kubernetes/pkg/api",
                 "k8s.io/kubernetes/pkg/api/events",
                 "k8s.io/kubernetes/pkg/api/legacyscheme",
-                "k8s.io/kubernetes/pkg/api/pod",
+                "k8s.io/kubernetes/pkg/apis/core/helper/pod",
                 "k8s.io/kubernetes/pkg/api/ref",
                 "k8s.io/kubernetes/pkg/api/resource",
                 "k8s.io/kubernetes/pkg/apis/core/helper/service",
